packages/hype_meter/requirements.txt
pymongo
python-dotenv
blockfrost-python
praw
pandas
scikit-learn
fastapi
uvicorn
numpy
